# Improve performance of EventQueueABM by better union splitting