before

```julia
julia> model = initialize_rps();

julia> @benchmark step!($model)
BenchmarkTools.Trial: 10000 samples with 197 evaluations.
 Range (min … max):  426.081 ns …  56.132 μs  ┊ GC (min … max): 0.00% … 98.59%
 Time  (median):     471.038 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   512.019 ns ± 745.854 ns  ┊ GC (mean ± σ):  5.40% ±  4.60%

         ▂▄▆▇█▇▆▄▃▁                                              
  ▁▁▂▃▄▅████████████▆▅▃▃▃▂▂▂▁▂▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▂▁ ▃
  426 ns           Histogram: frequency by time          641 ns <

 Memory estimate: 378 bytes, allocs estimate: 7.
```

after

```julia
julia> model = initialize_rps();

julia> @benchmark step!($model)
BenchmarkTools.Trial: 10000 samples with 207 evaluations.
 Range (min … max):  343.237 ns …  46.423 μs  ┊ GC (min … max): 0.00% … 98.19%
 Time  (median):     390.423 ns               ┊ GC (median):    0.00%
 Time  (mean ± σ):   409.583 ns ± 565.531 ns  ┊ GC (mean ± σ):  3.01% ±  2.53%

              ▂▃▅████▇▅▄▂▁                                       
  ▁▁▁▁▁▂▂▃▄▅▇█████████████▇▆▅▅▄▄▄▃▃▄▄▄▃▃▃▃▂▂▂▂▂▂▂▂▁▁▁▁▁▁▁▁▁▁▁▁▁ ▃
  343 ns           Histogram: frequency by time          495 ns <

 Memory estimate: 157 bytes, allocs estimate: 2.
```