@Datseris 2024-08-10
that's big!

can you give more info here? What are you doing? What is "union splitting" in this context?