@Tortar 2024-08-10
The key idea here in my opinion is understanding that tuples are not made for this task (at least currently) because when the access to a tuple is random and there are multiple types inside it, the tuple will not union split but just use a dynamic dispatch mechanism. When the access is fixed a tuple would be instead be better. A SizedVector is instead a good one because as a Vector it will union-split, at the same time we know the size of it, so we can use a fixed-size one.