@Tortar 2024-08-11
https://julialang.org/blog/2018/08/union-splitting/